Update set-orchestrator-type.md
The assumption of the initial statement here is that Docker EE is running exclusively on a host and Docker EE is not competing with other applications for resources. As well multi tenancy is assumed by this phrasing. As other use cases are possible like dedicated Docker clusters, Docker running on the same node where database applications are hosted or Docker EE is deployed on worker nodes with unknown state, e.g. shadow IT, things are different.

The statement not for production is use is too strict and does confuse users. Given the fact that an application owner or system administrator has to deal with out-of-memory and other resource constraints anyway all the time, it can be assumed that engineers know what they do and stating a informational or at least a warning would be more applicable.

Important is the case that service provider are forced to invest into additional consulting to mitigate this strict phrasing as customers are irritated seeing the restriction for production use. 

Most important this may impact the adoption of the product.